bridges/ri/pyproject.toml
[project]
name = "bridge-ri"
version = "0.1.0"
description = "A2A v0.3 ↔ v1.0 protocol bridge for ri_v10 (resource-intelligence)"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "a2a-sdk[http-server]>=1.0.2",
    # a2a-sdk 1.0.2 が FieldDescriptor.label に依存しており protobuf 7.x で削除されているため pin
    "protobuf>=6.33.0,<7",
    "httpx>=0.27",
    "starlette>=1.0",
    "uvicorn[standard]>=0.32",
    "sse-starlette>=3.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=1.0",
    "respx>=0.21",
]

[tool.hatch.build.targets.wheel]
packages = ["src/bridge_ri"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
